+++ donationpage/run.py
@@ -1,4 +1,4 @@
-from backend import app
+from donationpage import app
 
 if __name__ == '__main__':
     app.run('0.0.0.0',port=5000)
